+++ src/util/const.ts
@@ -58,6 +58,8 @@ export const EthersNetwork = {
   [Network.MATIC_MAINNET]: 'matic',
   [Network.MATIC_MUMBAI]: 'maticmum',
   [Network.MATIC_AMOY]: 'maticamoy',
+  [Network.SOLANA_MAINNET]: null, // Solana networks are not supported by ethers.js
+  [Network.SOLANA_DEVNET]: null, // Solana networks are not supported by ethers.js
   [Network.ASTAR_MAINNET]: 'astar-mainnet',
   [Network.POLYGONZKEVM_MAINNET]: 'polygonzkevm-mainnet',
   [Network.POLYGONZKEVM_TESTNET]: 'polygonzkevm-testnet',